+++ packages/accordion/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mhauens/accordion
 
+## 0.2.0
+
+### Minor Changes
+
+- remote styling
+
 ## 0.1.2
 
 ### Patch Changes