+++ packages/toggle/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @mhauens/toggle
 
+## 1.1.0
+
+### Minor Changes
+
+- remote styling
+
 ## 1.0.0
 
 ### Major Changes